minor bug fix

+++ controllers/contactController.js
@@ -53,7 +53,10 @@ const updateContact = asyncHandler(async (req, res) =>{
         res.status(404);
         throw new Error("Contact not found !");
     }
-    
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to update other user contacts !");
+    }
     const updateContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -71,6 +74,10 @@ const deleteContact = asyncHandler(async (req, res) =>{
     if (!contact){
         res.status(404);
         throw new Error("Contact not found !");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User dont have permission to delete other user contacts !");
     }
     await Contact.deleteOne();
     res.status(200).json(contact);
